Address Markov chain review feedback
