+++ packages/quicktype-core/src/support/Support.ts
@@ -1,6 +1,3 @@
-import * as YAML from "yaml";
-
-import type { JSONSchema } from "../input/JSONSchemaStore.js";
 import { messageError } from "../Messages.js";
 
 export interface StringMap {
@@ -133,35 +130,6 @@ export function errorMessage(e: unknown): string {
     }
 
     return (e as { toString: () => string }).toString();
-}
-
-export function parseJSON(
-    text: string,
-    description: string,
-    address = "<unknown>",
-): JSONSchema | undefined {
-    try {
-        // https://gist.github.com/pbakondy/f5045eff725193dad9c7
-        if (text.charCodeAt(0) === 0xfeff) {
-            text = text.slice(1);
-        }
-
-        return YAML.parse(text);
-    } catch (e) {
-        let message: string;
-
-        if (e instanceof SyntaxError) {
-            message = e.message;
-        } else {
-            message = `Unknown exception ${e}`;
-        }
-
-        return messageError("MiscJSONParseError", {
-            description,
-            address,
-            message,
-        });
-    }
 }
 
 export function indentationString(level: number): string {
